requirements.txt
asgiref==3.2.3
astroid==2.2.5
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.1
dj-database-url==0.5.0
Django==2.1.7
django-bootstrap4==0.0.8
django-cors-headers==3.2.1
django-crispy-forms==1.7.2
django-filter==2.2.0
django-heroku==0.3.1
django-tinymce==2.8.0
djangorestframework==3.11.0
gunicorn==19.9.0
idna==2.8
isort==4.3.17
lazy-object-proxy==1.3.1
Markdown==3.2
mccabe==0.6.1
psycopg2==2.8.3
psycopg2-binary==2.8.4
pylint==2.3.1
python-dotenv==0.10.3
pytz==2018.9
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.4
urllib3==1.25.8
whitenoise==4.1.2
wrapt==1.11.1
